# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Tutorium Backend system.",
        "title": "Tutorium Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en."
        },
        "version": "1.0"
    },
    "paths": {
...
        "/banlearners": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GetBanLearners returns a list of all ban records",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BanLearners"
                ],
                "summary": "Get all banned learners",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BanDetailsLearnerDoc"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "CreateBanLearner creates a new ban record for a learner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BanLearners"
                ],
                "summary": "Create a new banned learner record",
                "parameters": [
                    {
                        "description": "Ban Learner Payload",
                        "name": "banlearner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BanDetailsLearnerDoc"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.BanDetailsLearnerDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/banlearners/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GetBanLearner returns a single ban record by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BanLearners"
                ],
                "summary": "Get a banned learner by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ban Learner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BanDetailsLearnerDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "UpdateBanLearner modifies an existing ban record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BanLearners"
                ],
                "summary": "Update a banned learner record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ban Learner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated ban record",
                        "name": "banlearner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BanDetailsLearnerDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BanDetailsLearnerDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "DeleteBanLearner deletes a ban record by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BanLearners"
                ],
                "summary": "Delete a banned learner record",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ban Learner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted ban learner",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/learners": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GetLearners retrieves all Learner records",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learners"
                ],
                "summary": "List all learners",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.LearnerDoc"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "CreateLearner creates a new Learner record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learners"
                ],
                "summary": "Create a new learner",
                "parameters": [
                    {
                        "description": "Learner payload",
                        "name": "learner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LearnerDoc"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.LearnerDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/learners/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GetLearner retrieves a single Learner by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learners"
                ],
                "summary": "Get learner by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Learner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LearnerDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Learner not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "DeleteLearner removes a Learner record by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Learners"
                ],
                "summary": "Delete a learner by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Learner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted Learner",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Learner not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BanDetailsLearnerDoc": {
            "type": "object",
            "properties": {
                "ban_description": {
                    "type": "string",
                    "example": "Spamming inappropriate content"
                },
                "ban_end": {
                    "type": "string",
                    "example": "2025-08-30T12:00:00Z"
                },
                "ban_start": {
                    "type": "string",
                    "example": "2025-08-20T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "learner_id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "models.LearnerDoc": {
            "type": "object",
            "properties": {
                "ban_count": {
                    "type": "integer",
                    "example": 1
                },
                "flag_count": {
                    "type": "integer",
                    "example": 3
                },
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "user_id": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token.`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}